+++ internal/config/load.go
@@ -1,4 +1,4 @@
-package am
+package config
 
 import (
 	"encoding/json"
