+++ internal/config/source_tracking_simple_test.go
@@ -1,4 +1,4 @@
-package am
+package config
 
 import (
 	"os"
