+++ internal/config/watcher.go
@@ -1,4 +1,4 @@
-package am
+package config
 
 import (
 	"path/filepath"
